// src/graphql/shopify/storefront/graphql.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K]
}
// ...
export type Cart = HasMetafields &
  Node & {
    __typename?: 'Cart'
    /** The gift cards that have been applied to the cart. */
    appliedGiftCards: Array<AppliedGiftCard>
    /** An attribute associated with the cart. */
    attribute?: Maybe<Attribute>
    /** The attributes associated with the cart. Attributes are represented as key-value pairs. */
    attributes: Array<Attribute>
    /** Information about the buyer that's interacting with the cart. */
    buyerIdentity: CartBuyerIdentity
    /** The URL of the checkout for the cart. */
    checkoutUrl: Scalars['URL']['output']
    /** The estimated costs that the buyer will pay at checkout. The costs are subject to change and changes will be reflected at checkout. The `cost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing). */
    cost: CartCost
    /** The date and time when the cart was created. */
    createdAt: Scalars['DateTime']['output']
    /**
     * The delivery groups available for the cart, based on the buyer identity default
     * delivery address preference or the default address of the logged-in customer.
     *
     */
    deliveryGroups: CartDeliveryGroupConnection
    /** The discounts that have been applied to the entire cart. */
    discountAllocations: Array<CartDiscountAllocation>
    /** The case-insensitive discount codes that the customer added at checkout. */
    discountCodes: Array<CartDiscountCode>
    /**
     * The estimated costs that the buyer will pay at checkout. The estimated costs are subject to change and changes will be reflected at checkout. The `estimatedCost` field uses the `buyerIdentity` field to determine [international pricing](https://shopify.dev/custom-storefronts/internationalization/international-pricing).
     * @deprecated Use `cost` instead.
     */
    estimatedCost: CartEstimatedCost
    /** A globally-unique ID. */
    id: Scalars['ID']['output']
    /** A list of lines containing information about the items the customer intends to purchase. */
    lines: BaseCartLineConnection
    /** Returns a metafield found by namespace and key. */
    metafield?: Maybe<Metafield>
    /** The metafields associated with the resource matching the supplied list of namespaces and keys. */
    metafields: Array<Maybe<Metafield>>
    /** A note that's associated with the cart. For example, the note can be a personalized message to the buyer. */
    note?: Maybe<Scalars['String']['output']>
    /** The total number of items in the cart. */
    totalQuantity: Scalars['Int']['output']
    /** The date and time when the cart was updated. */
    updatedAt: Scalars['DateTime']['output']
  }
// ...
export type ProductConnection = {
  __typename?: 'ProductConnection'
  /** A list of edges. */
  edges: Array<ProductEdge>
  /** A list of available filters. */
  filters: Array<Filter>
  /** A list of the nodes contained in ProductEdge. */
  nodes: Array<Product>
  /** Information to aid in pagination. */
  pageInfo: PageInfo
}

/**
 * An auto-generated type which holds one Product and a cursor during pagination.
 *
 */
export type ProductEdge = {
  __typename?: 'ProductEdge'
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output']
  /** The item at the end of ProductEdge. */
  node: Product
}
// ...
export type ProductVariant = HasMetafields &
  Node & {
    __typename?: 'ProductVariant'
    /** Indicates if the product variant is available for sale. */
    availableForSale: Scalars['Boolean']['output']
    /** The barcode (for example, ISBN, UPC, or GTIN) associated with the variant. */
    barcode?: Maybe<Scalars['String']['output']>
    /** The compare at price of the variant. This can be used to mark a variant as on sale, when `compareAtPrice` is higher than `price`. */
    compareAtPrice?: Maybe<MoneyV2>
    /**
     * The compare at price of the variant. This can be used to mark a variant as on sale, when `compareAtPriceV2` is higher than `priceV2`.
     * @deprecated Use `compareAtPrice` instead.
     */
    compareAtPriceV2?: Maybe<MoneyV2>
    /**
     * List of bundles components included in the variant considering only fixed bundles.
     *
     */
    components: ProductVariantComponentConnection
    /** Whether a product is out of stock but still available for purchase (used for backorders). */
    currentlyNotInStock: Scalars['Boolean']['output']
    /**
     * List of bundles that include this variant considering only fixed bundles.
     *
     */
    groupedBy: ProductVariantConnection
    /** A globally-unique ID. */
    id: Scalars['ID']['output']
    /** Image associated with the product variant. This field falls back to the product image if no image is available. */
    image?: Maybe<Image>
    /** Returns a metafield found by namespace and key. */
    metafield?: Maybe<Metafield>
    /** The metafields associated with the resource matching the supplied list of namespaces and keys. */
    metafields: Array<Maybe<Metafield>>
    /** The product variant’s price. */
    price: MoneyV2
    /**
     * The product variant’s price.
     * @deprecated Use `price` instead.
     */
    priceV2: MoneyV2
    /** The product object that the product variant belongs to. */
    product: Product
    /** The total sellable quantity of the variant for online sales channels. */
    quantityAvailable?: Maybe<Scalars['Int']['output']>
    /** A list of quantity breaks for the product variant. */
    quantityPriceBreaks: QuantityPriceBreakConnection
    /** The quantity rule for the product variant in a given context. */
    quantityRule: QuantityRule
    /**
     * Whether a product variant requires components. The default value is `false`.
     * If `true`, then the product variant can only be purchased as a parent bundle with components.
     *
     */
    requiresComponents: Scalars['Boolean']['output']
    /** Whether a customer needs to provide a shipping address when placing an order for the product variant. */
    requiresShipping: Scalars['Boolean']['output']
    /** List of product options applied to the variant. */
    selectedOptions: Array<SelectedOption>
    /** Represents an association between a variant and a selling plan. Selling plan allocations describe which selling plans are available for each variant, and what their impact is on pricing. */
    sellingPlanAllocations: SellingPlanAllocationConnection
    /** The Shop Pay Installments pricing information for the product variant. */
    shopPayInstallmentsPricing?: Maybe<ShopPayInstallmentsProductVariantPricing>
    /** The SKU (stock keeping unit) associated with the variant. */
    sku?: Maybe<Scalars['String']['output']>
    /** The in-store pickup availability of this variant by location. */
    storeAvailability: StoreAvailabilityConnection
    /** Whether tax is charged when the product variant is sold. */
    taxable: Scalars['Boolean']['output']
    /** The product variant’s title. */
    title: Scalars['String']['output']
    /** The unit price value for the variant based on the variant's measurement. */
    unitPrice?: Maybe<MoneyV2>
    /** The unit price measurement for the variant. */
    unitPriceMeasurement?: Maybe<UnitPriceMeasurement>
    /** The weight of the product variant in the unit system specified with `weight_unit`. */
    weight?: Maybe<Scalars['Float']['output']>
    /** Unit of measurement for weight. */
    weightUnit: WeightUnit
  }
// ...
export type ProductsQueryVariables = Exact<{ [key: string]: never }>

export type ProductsQuery = {
  __typename?: 'QueryRoot'
  products: {
    __typename?: 'ProductConnection'
    edges: Array<{
      __typename?: 'ProductEdge'
      cursor: string
      node: {
        __typename?: 'Product'
        id: string
        title: string
        description: string
        handle: string
        variants: {
          __typename?: 'ProductVariantConnection'
          edges: Array<{
            __typename?: 'ProductVariantEdge'
            cursor: string
            node: {
              __typename?: 'ProductVariant'
              availableForSale: boolean
              id: string
              title: string
              price: {
                __typename?: 'MoneyV2'
                amount: any
                currencyCode: CurrencyCode
              }
            }
          }>
        }
      }
    }>
  }
}

export type CartMutationVariables = Exact<{
  merchandiseId: Scalars['ID']['input']
  quantity: Scalars['Int']['input']
}>

export type CartMutation = {
  __typename?: 'Mutation'
  cartCreate?: {
    __typename?: 'CartCreatePayload'
    cart?: {
      __typename?: 'Cart'
      id: string
      createdAt: any
      updatedAt: any
      checkoutUrl: any
      lines: {
        __typename?: 'BaseCartLineConnection'
        edges: Array<{
          __typename?: 'BaseCartLineEdge'
          node:
            | {
                __typename?: 'CartLine'
                id: string
                merchandise: { __typename?: 'ProductVariant'; id: string }
              }
            | {
                __typename?: 'ComponentizableCartLine'
                id: string
                merchandise: { __typename?: 'ProductVariant'; id: string }
              }
        }>
      }
      attributes: Array<{
        __typename?: 'Attribute'
        key: string
        value?: string | null
      }>
      cost: {
        __typename?: 'CartCost'
        totalAmount: {
          __typename?: 'MoneyV2'
          amount: any
          currencyCode: CurrencyCode
        }
        subtotalAmount: {
          __typename?: 'MoneyV2'
          amount: any
          currencyCode: CurrencyCode
        }
        totalTaxAmount?: {
          __typename?: 'MoneyV2'
          amount: any
          currencyCode: CurrencyCode
        } | null
        totalDutyAmount?: {
          __typename?: 'MoneyV2'
          amount: any
          currencyCode: CurrencyCode
        } | null
      }
    } | null
  } | null
}

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType']

  constructor(
    private value: string,
    public __meta__?: Record<string, any> | undefined,
  ) {
    super(value)
  }

  toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value
  }
}

export const ProductsDocument = new TypedDocumentString(`
    query Products {
  products(first: 10) {
    edges {
      cursor
      node {
        id
        title
        description
        handle
        variants(first: 10) {
          edges {
            cursor
            node {
              availableForSale
              id
              title
              price {
                amount
                currencyCode
              }
            }
          }
        }
      }
    }
  }
}
    `) as unknown as TypedDocumentString<ProductsQuery, ProductsQueryVariables>
export const CartDocument = new TypedDocumentString(`
    mutation Cart($merchandiseId: ID!, $quantity: Int!) {
  cartCreate(
    input: {lines: [{merchandiseId: $merchandiseId, quantity: $quantity}]}
  ) {
    cart {
      id
      createdAt
      updatedAt
      checkoutUrl
      lines(first: 10) {
        edges {
          node {
            id
            merchandise {
              ... on ProductVariant {
                id
              }
            }
          }
        }
      }
      attributes {
        key
        value
      }
      cost {
        totalAmount {
          amount
          currencyCode
        }
        subtotalAmount {
          amount
          currencyCode
        }
        totalTaxAmount {
          amount
          currencyCode
        }
        totalDutyAmount {
          amount
          currencyCode
        }
      }
    }
  }
}
    `) as unknown as TypedDocumentString<CartMutation, CartMutationVariables>